v3 proto: renamed vindex files to match what they define
