+++ go/vt/vtgate/vindexes/lookup_hash_test.go
