+++ go/vt/vtgate/vindexes/numeric.go
