+++ go/vt/vtgate/vindexes/numeric_test.go
